Delete old code, create the data file from this branch instead
