--- .eslintrc.json
@@ -1,42 +1,0 @@
-{
-    "parser": "babel-eslint",
-    "plugins": [
-        "prettier"
-    ],
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es6": true,
-        "node": true,
-        "jasmine": true,
-        "jest": true
-    },
-    "parserOptions": {
-        "ecmaVersion": 2017,
-        "ecmaFeatures": {},
-        "sourceType": "module",
-        "experimentalObjectRestSpread": true
-    },
-    "rules": {
-        "prettier/prettier": ["error", {
-            "printWidth": 200,
-            "tabWidth": 4,
-            "bracketSpacing": false
-        }],
-        "max-len": ["warn", {
-            "code": 200,
-            "comments": 80,
-            "ignoreTrailingComments": true
-        }],
-        "@typescript-eslint/camelcase": "off",
-        "linebreak-style": ["error","unix"],
-        "camelcase": "off",
-        "no-const-assign": "error",
-        "no-this-before-super": "error",
-        "no-undef": "error",
-        "no-unreachable": "error",
-        "no-unused-vars": "error",
-        "constructor-super": "error",
-        "valid-typeof": "error"
-    }
-}
