+++ README.md
@@ -1,4 +1,7 @@
 NYPD CCRB analytics from various sources
 
-* [New York Civil Liberties Union, Github](https://github.com/new-york-civil-liberties-union)
-* [NYPD Member of Service Histories, nyc.gov](https://www1.nyc.gov/site/ccrb/policy/MOS-records.page)
+The script here gathers the data from the [NYCLU updated data](https://github.com/new-york-civil-liberties-union/NYPD-Misconduct-Complaint-Database-Updated), and outputs a file `nypdccrb.arrow`.
+We use this file in the `gh-pages` branch to serve it.
+
+-   [New York Civil Liberties Union, Github](https://github.com/new-york-civil-liberties-union)
+-   [NYPD Member of Service Histories, nyc.gov](https://www1.nyc.gov/site/ccrb/policy/MOS-records.page)
