--- src/dark.css
@@ -1,10 +1,0 @@
-
-.dark button {
-    background-color: inherit;
-    color: white;
-    border: 1px solid rgba(255,255,255,0.2);
-}
-
-.dark #buttons {
-    background-color: #242526;
-}
