--- src/index.css
@@ -1,52 +1,0 @@
-
-perspective-workspace { 
-    position: absolute; 
-    top: 0; 
-    left: 0; 
-    right: 0; 
-    bottom: 0; 
-}
-
-#buttons > * {
-    padding: 6px;
-    margin: 5px;
-}
-
-button {
-    background-color: inherit;
-    cursor: pointer;
-}
-
-button:hover {
-    border: 1px solid #00f !important;
-}
-
-body {
-    overflow: hidden;
-}
-
-#message {
-    font-family: monospace;
-}
-
-#buttons {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    height: 36px;
-}
-
-#workspace {
-    height: auto;
-    position: absolute;
-    top: 36px;
-    right: 0;
-    bottom: 0;
-    left: 0;
-}
-
-a {
-    text-transform: uppercase;
-    font-size: 10px;
-}
