--- src/light.css
@@ -1,9 +1,0 @@
-
-button {
-    background-color: inherit;
-    border: 1px solid rgba(0,0,0,0.5);
-}
-
-#buttons {
-    background-color: #e1e6ea;
-}
